requirements.txt
matplotlib==3.5.1
natsort==8.1.0
numpy==1.22.0
pandas==1.3.5
Pillow==9.0.1
scikit_learn==1.0.2
torch==1.11.0a0+bfe5ad2
torchinfo==1.6.3
torchvision==0.12.0a0
tqdm==4.62.3
